pyproject.toml
[project]
name = "as5"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt>=4.2.0",
    "ddgs>=9.11.3",
    "fastapi>=0.135.1",
    "litellm>=1.82.1",
    "openai>=2.26.0",
    "openai-agents>=0.12.0",
    "passlib>=1.7.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "scalar-fastapi>=1.8.0",
    "sqlmodel>=0.0.37",
    "streamlit>=1.44.0",
]
